Cambios de username en email

+++ routes/api.php
@@ -12,26 +12,27 @@ Route::post('/login', [AuthController::class, 'login'])->middleware('throttle:5,
 Route::post('/register/create', [AuthController::class, 'create_register'])->middleware('throttle:5,1'); // Esto es para loguearse
 Route::get('/register/complete/data/information', [AuthController::class, 'sendCompleteDataEmail']); // Esto es para enviar el correo
 Route::post('/verify-email/{token}', [AuthController::class, 'verifyEmail'])->middleware('throttle:5,1');
-Route::get('/users' , [PostsController::class, 'show']); // Esto muestra los usuarios
-Route::middleware(['auth:sanctum'])->get('/posting/{id}', [PostsController::class, 'postsPaginacion']); // Esto muestra los posts de los usuarios
-Route::middleware(['auth:sanctum'])->get('/post/{id}', [PostsController::class, 'showComment']); // Esto muestra los comentarios del post
-Route::middleware(['auth:sanctum'])->get('/posts/{postId}/latest-comment', [CommentController::class, 'showLastCommentModal']); // Esto muestra el último comentario
-Route::middleware(['auth:sanctum'])->post('/posts/comments', [CommentController::class, 'create']); // Esto crea un comentario
-Route::middleware(['auth:sanctum'])->post('/posts/reply', [CommentReplyController::class, 'create']); // Esto crea una respuesta a un comentario
-Route::middleware(['auth:sanctum'])->post('/post/{postId}/{userId}/like', [LikeController::class, 'likePost']); // Esto da like a un post
-Route::middleware(['auth:sanctum'])->delete('/post/{postId}/{userId}/like', [LikeController::class, 'unlikePost']); // Esto quita el like a un post
+Route::get('/users', [PostsController::class, 'show']); // Esto muestra los usuarios
+Route::middleware(['auth:sanctum'])->group(function () {
+    Route::get('/posting/{id}', [PostsController::class, 'postsPaginacion']); // Esto muestra los posts de los usuarios
+    Route::get('/post/{id}', [PostsController::class, 'showComment']); // Esto muestra los comentarios del post
+    Route::get('/posts/{postId}/latest-comment', [CommentController::class, 'showLastCommentModal']); // Esto muestra el último comentario
+    Route::post('/posts/comments', [CommentController::class, 'create']); // Esto crea un comentario
+    Route::post('/posts/reply', [CommentReplyController::class, 'create']); // Esto crea una respuesta a un comentario
+    Route::post('/post/{postId}/{userId}/like', [LikeController::class, 'likePost']); // Esto da like a un post
+    Route::delete('/post/{postId}/{userId}/like', [LikeController::class, 'unlikePost']); // Esto quita el like a un post
+});
 
 Route::options('{any}', function (Request $request) {
     // Orígenes permitidos
-    $allowedOrigins = ['http://localhost:5173', 'http://tusitio.com'];
+    $allowedOrigins = ['http://localhost:5173', 'https://api-foro-social.myhostingspace.net'];
 
-    // Verifica si el origen de la solicitud está permitido
-    $origin = $request->headers->get('Origin');
+    $origin = $request->headers->get('Origin'); // Verifica si el origen de la solicitud está permitido
     if (in_array($origin, $allowedOrigins)) {
         return response()->json([], 204)
-                        ->header('Access-Control-Allow-Origin', $origin)
-                        ->header('Access-Control-Allow-Methods', 'GET, POST, OPTIONS, PUT, DELETE')
-                        ->header('Access-Control-Allow-Headers', 'Content-Type, Authorization');
+            ->header('Access-Control-Allow-Origin', $origin)
+            ->header('Access-Control-Allow-Methods', 'GET, POST, OPTIONS, PUT, DELETE')
+            ->header('Access-Control-Allow-Headers', 'Content-Type, Authorization');
     }
 
     // Si el origen no está permitido, devuelve un error
@@ -41,20 +42,18 @@ Route::options('{any}', function (Request $request) {
 // sactum es un middleware de autenticación y throttle (code 429) es un middleware de limitación de peticiones (5 x min).
 // cada dirección IP tendrá su propio límite de 5 intentos por minuto.
 
+
+
+
 /*
-Route::middleware(['auth:sanctum'])->group(function () {
-    Route::get('/users', [PostsController::class, 'show']);
-    Route::get('/posting/{id}', [PostsController::class, 'postsPaginacion']);
-    Route::get('/post/{id}', [PostsController::class, 'showComment']);
-    Route::get('/posts/{postId}/latest-comment', [CommentController::class, 'showLastCommentModal']);
-    Route::post('/posts/comments', [CommentController::class, 'create']);
-    Route::post('/posts/reply', [CommentReplyController::class, 'create']);
-    Route::post('/post/{postId}/{userId}/like', [LikeController::class, 'likePost']);
-    Route::delete('/post/{postId}/{userId}/like', [LikeController::class, 'unlikePost']);
-});
-*/
-
-
+Route::middleware(['auth:sanctum'])->get('/posting/{id}', [PostsController::class, 'postsPaginacion']); // Esto muestra los posts de los usuarios
+Route::middleware(['auth:sanctum'])->get('/post/{id}', [PostsController::class, 'showComment']); // Esto muestra los comentarios del post
+Route::middleware(['auth:sanctum'])->get('/posts/{postId}/latest-comment', [CommentController::class, 'showLastCommentModal']); // Esto muestra el último comentario
+Route::middleware(['auth:sanctum'])->post('/posts/comments', [CommentController::class, 'create']); // Esto crea un comentario
+Route::middleware(['auth:sanctum'])->post('/posts/reply', [CommentReplyController::class, 'create']); // Esto crea una respuesta a un comentario
+Route::middleware(['auth:sanctum'])->post('/post/{postId}/{userId}/like', [LikeController::class, 'likePost']); // Esto da like a un post
+Route::middleware(['auth:sanctum'])->delete('/post/{postId}/{userId}/like', [LikeController::class, 'unlikePost']); // Esto quita el like a un post
+/*
